Ensure InteractableContainerBlock opens the correct block
Log when a container is behind blocks
Change cs_info to print decimals for entity containers and the correct
block position for block containers

+++ fabric-client/src/main/java/org/samo_lego/clientstorage/fabric_client/storage/InteractableContainer.java
@@ -85,10 +85,10 @@ public interface InteractableContainer extends Container {
      */
     default String cs_info() {
         // Get position
-        final String position = String.format("(%d, %d, %d)",
-                (int) this.cs_position().x(),
-                (int) this.cs_position().y(),
-                (int) this.cs_position().z());
+        final String position = String.format("(%.2f, %.2f, %.2f)",
+                this.cs_position().x(),
+                this.cs_position().y(),
+                this.cs_position().z());
         return String.format("%s @ %s [%d slots]", this.cs_getName().getString(), position, this.getContainerSize());
     }
 }
